implement key event stream parsing
